Include Maven SNAPSHOT pre-release example
The Maven build system uses the keyword "SNAPSHOT" to indicate pre-release builds. Their definition and use fits nicely with the semver pre-release definition so it would be useful to be shown here as an example. Worth mentioning is that the "SNAPSHOT" naming convention is sometimes used by non-Maven projects too.

"you will often see the SNAPSHOT designator in a version, which indicates that a project is in a state of development"
http://maven.apache.org/guides/getting-started