Switched from Lexical Sort to Natural Sort
This change better first with people’s expectations. This way
1.0.0-beta2 has a lower precedence than 1.0.0-beta11, because “2” is
smaller than “11”- a point missed by lexical sorting. It also solves
the alphabetical issue  because natural sorting is essentially case
insensitive (it may or may not make sense to specify that two versions
of the same exact name but different cases is frowned upon, in which
case it’s case sensitive for our purposes).