Changed MUST to MAY for resetting Minor and Patch when incrementing Major
There is no need to reset the Minor version number when incrementing the Major, as the new Major version may include the same functionality as the previous Major version.

Patch version numbers need only be reset when the functionality changes. I.e. when the Minor Version is incremented.

Resetting the Minor version to Zero looses the Minor Version Number's helpfulness in seeing at-a-glance how many new feature versions have been developed.

This makes the following legal:

    1.2.3 -> 2.2.3     // A compatibility-breaking change was introduced, but no patches or new functionality
    1.2.3 -> 2.2.4     // A compatibility-breaking patch was introduced, but no new functionality
    1.2.3 -> 2.3.0     // A compatibility-breaking feature was introduced
    1.2.3 -> 2.0.0     // Any of the above combinations, but you can still reset both if you like!